chore: move classes and add tests
