+++ components/engine/engine-camel/src/main/java/org/eclipse/dirigible/components/engine/camel/processor/CamelDirigibleRequestHandlerMapping.java
@@ -5,8 +5,6 @@ import org.apache.camel.component.platform.http.PlatformHttpComponent;
 import org.apache.camel.component.platform.http.spi.PlatformHttpEngine;
 import org.apache.camel.component.platform.http.springboot.CamelRequestHandlerMapping;
 
-import java.nio.file.Path;
-
 public class CamelDirigibleRequestHandlerMapping extends CamelRequestHandlerMapping {
 
     public CamelDirigibleRequestHandlerMapping(PlatformHttpComponent component, PlatformHttpEngine engine) {
@@ -15,9 +13,7 @@ public class CamelDirigibleRequestHandlerMapping extends CamelRequestHandlerMapp
 
     @Override
     public void registerHttpEndpoint(HttpEndpointModel model) {
-        var patchedUri = Path.of("/services/integrations", model.getUri())
-                             .toString();
-        var patchedModel = new HttpEndpointModel(patchedUri, model.getVerbs(), model.getConsumer());
+        var patchedModel = DirigibleHttpEndpointModel.from(model);
         super.registerHttpEndpoint(patchedModel);
     }
 }
